+++ src/Transaction/Transaction.php
@@ -33,6 +33,21 @@ class Transaction
     /** @var array|null */
     protected $processorInformation;
 
+    /** @var null|\DateTime */
+    protected $createdAt;
+
+    /** @var string|null */
+    protected $origin;
+
+    /** @var string|null */
+    protected $destination;
+
+    /** @var string|null */
+    protected $currency;
+
+    /** @var int|null */
+    protected $orderId;
+
     public function __construct(array $data)
     {
         $this->id = $data['transactionId'];
@@ -42,6 +57,17 @@ class Transaction
         $this->amount = $data['amount'];
         $this->processorName = $data['processorName'];
         $this->processorInformation = $data['processorInformation'];
+        if (\array_key_exists('transactionDate', $data) === true
+            && $data['transactionDate'] !== null
+        ) {
+            $this->createdAt = new \DateTime($data['transactionDate']);
+        } else {
+            $this->createdAt = null;
+        }
+        $this->origin = $data['origin'] ?? null;
+        $this->destination = $data['destination'] ?? null;
+        $this->currency = $data['currency'] ?? null;
+        $this->orderId = $data['order_id'] ?? null;
     }
 
     public function getId(): string
@@ -77,5 +103,30 @@ class Transaction
     public function getProcessorInformation(): ?array
     {
         return $this->processorInformation;
+    }
+
+    public function getCreatedAt(): ?\DateTime
+    {
+        return $this->createdAt;
+    }
+
+    public function getOrigin(): ?string
+    {
+        return $this->origin;
+    }
+
+    public function getDestination(): ?string
+    {
+        return $this->destination;
+    }
+
+    public function getCurrency(): ?string
+    {
+        return $this->currency;
+    }
+
+    public function getOrderId(): ?int
+    {
+        return $this->orderId;
     }
 }